requirements.txt
pyTelegramBotAPI
gpt-2-simple
python-dotenv
redis
tensorflow==1.15